Cell text style

Functions to get or set the cell text style property of huxtable cells.

bold(ht)
bold(ht) <- value
set_bold(ht, row, col, value, byrow = FALSE)
map_bold(ht, row, col, fn)

italic(ht)
italic(ht) <- value
set_italic(ht, row, col, value, byrow = FALSE)
map_italic(ht, row, col, fn)

Arguments

ht
A huxtable.

value
A logical vector or matrix. TRUE for bold/italic.

Set to NA to reset to the default, which is FALSE.

row
A row specifier. See rowspecs for details.

col
An optional column specifier.

fn
A mapping function. See mapping-functions for details.

byrow
Deprecated. Use by_cols() instead.

Value

For bold, the bold property. For set_bold and map_bold, the modified huxtable.

Similarly for italic and friends.

See also

Examples

bold(jams) <-  TRUE
bold(jams)

#>   Type price   V1   V1   V1   V1
#> 1 TRUE  TRUE TRUE TRUE TRUE TRUE
#> 2 TRUE  TRUE TRUE TRUE TRUE TRUE
#> 3 TRUE  TRUE TRUE TRUE TRUE TRUE
#> 4 TRUE  TRUE TRUE TRUE TRUE TRUE

set_bold(jams, TRUE)

#>   Type         Price    Sugar    Sugar    Sugar    Sugar  
#>   Strawberry    1.90   60.00%   60.00%   60.00%   60.00%  
#>   Raspberry     2.10   60.00%   60.00%   60.00%   60.00%  
#>   Plum          1.80   60.00%   60.00%   60.00%   60.00%  
#> 
#> Column names: Type, price, V1, V1, V1, V1

set_bold(jams,
      2:3, 1, TRUE)

#>   Type         Price    Sugar    Sugar    Sugar    Sugar  
#>   Strawberry    1.90   60.00%   60.00%   60.00%   60.00%  
#>   Raspberry     2.10   60.00%   60.00%   60.00%   60.00%  
#>   Plum          1.80   60.00%   60.00%   60.00%   60.00%  
#> 
#> Column names: Type, price, V1, V1, V1, V1

map_bold(jams,
      by_rows(TRUE, FALSE))

#>   Type         Price    Sugar    Sugar    Sugar    Sugar  
#>   Strawberry    1.90   60.00%   60.00%   60.00%   60.00%  
#>   Raspberry     2.10   60.00%   60.00%   60.00%   60.00%  
#>   Plum          1.80   60.00%   60.00%   60.00%   60.00%  
#> 
#> Column names: Type, price, V1, V1, V1, V1