by_values {huxtable}	R Documentation

Map specific cell values to cell properties

Description

Use by_values() to set properties for cells with specific, pre-determined contents.

Usage

by_values(..., ignore_na = TRUE)

Arguments

...
Name-value pairs like name = value. Cells where contents are equal to name will have the property set to value. If there is a single unnamed argument, this is the default value for unmatched cells. More than one unnamed argument is an error.

ignore_na
If TRUE, NA values in the result will be left unchanged from their previous values. Otherwise, NA normally resets to the default.

Value

A function for use in map_*** functions.

See Also

mapping-functions

Other mapping functions: by_cases(), by_colorspace(), by_function(), by_quantiles(), by_ranges(), by_regex(), by_rows()

Examples

ht <- hux(letters[1:3])
map_background_color(
  ht,
  by_values(a = "red", c = "yellow")
)
map_background_color(
  ht,
  by_values(a = "red", c = "yellow", "green")
)

[Package huxtable version 5.8.0 Index]