col_width {huxtable}	R Documentation

Set the width of table columns

Description

Numeric column widths are treated as proportions of the table width. Character widths must be valid CSS or LaTeX dimensions.

Usage

col_width(ht)

col_width(ht) <- value

set_col_width(ht, col, value)

Arguments

ht
A huxtable.

value
Numeric or character vector. Set to NA to reset to the default, which is NA.

col
A column specifier. See rowspecs for details.

Details

In LaTeX, if you specify a column width, but set wrap to FALSE and have cells which overrun, then you may have problems with table position and with background colours in other cells. The workaround is to adjust the width, so that your cells no longer overrun.

Value

col_width() returns the col_width property. set_col_width() returns the modified huxtable.

See Also

Other table measurements: height(), row_height(), width()

Examples

col_width(jams) <- c(.2, .8)
col_width(jams)

jams$Notes <- c(
  "Notes",
  "This year's finest", "", ""
)
jams
set_col_width(jams, c(.4, .5, .1))

[Package huxtable version 5.8.0 Index]