Font size

Functions to get or set the font size property of huxtable cells.

font_size(ht)
font_size(ht) <- value
set_font_size(ht, row, col, value, byrow = FALSE)
map_font_size(ht, row, col, fn)

Arguments

ht
A huxtable.

value
A numeric vector. This sets the font size in points.

Set to NA to reset to the default, which is NA.

row
A row specifier. See rowspecs for details.

col
An optional column specifier.

fn
A mapping function. See mapping-functions for details.

byrow
Deprecated. Use by_cols() instead.

Value

For font_size, the font_size property. For set_font_size and map_font_size, the modified huxtable.

See also

Examples

orig <- font_size(jams)
font_size(jams) <-  14
font_size(jams)

#>   Type price
#> 1   14    14
#> 2   14    14
#> 3   14    14
#> 4   14    14

font_size(jams) <- orig

jams2 <- set_font_size(jams,
      14)
font_size(jams2)

#>   Type price
#> 1   14    14
#> 2   14    14
#> 3   14    14
#> 4   14    14

jams3 <- set_font_size(jams,
      2:3, 1, 14)
font_size(jams3)

#>   Type price
#> 1   NA    NA
#> 2   14    NA
#> 3   14    NA
#> 4   NA    NA

jams4 <- map_font_size(jams,
      by_rows(
        14,
        12)
      )
font_size(jams4)

#>   Type price
#> 1   14    14
#> 2   12    12
#> 3   14    14
#> 4   12    12