guess_knitr_output_format {huxtable}	R Documentation

Guess knitr output format

Description

Convenience function which tries to guess the ultimate output from knitr and rmarkdown.

Usage

guess_knitr_output_format()

Value

"html", "latex", "typst", or something else. If we are not in a knitr document, returns an empty string.

Examples

## Not run: 
# in a knitr document
guess_knitr_output_format()

## End(Not run)

[Package huxtable version 5.8.0 Index]