height {huxtable}	R Documentation

Set the table height

Description

height() sets the height of the entire table, while row_height() sets the height of individual rows. A numeric height is treated as a proportion of the containing block (HTML) or \textheight (LaTeX). A character height must be a valid CSS or LaTeX dimension.

Usage

height(ht)

height(ht) <- value

set_height(ht, value)

Arguments

ht
A huxtable.

value
A number or string. Set to NA to reset to the default, which is NA_real_.

Value

property() returns the property value(s). set_property() and map_property() return the modified huxtable.

See Also

Other table measurements: col_width(), row_height(), width()

Examples

set_height(jams, 0.4)

[Package huxtable version 5.8.0 Index]