Frequently Asked Questions, including how to get help

A FAQ of common issues.

Details

I get a LaTeX error when I try to compile my document!

Have you installed the LaTeX packages you need? LaTeX packages are different from R packages. Run check_latex_dependencies() to find out if you are missing any. Then install them using your system's LaTeX management application. Or you can try install_latex_dependencies().

In some rmarkdown and LaTeX formats, you also need to add LaTeX dependencies manually. Run report_latex_dependencies() and add the output to your LaTeX preamble, or in Rmarkdown formats, add it to the rmarkdown header like this:

Huxtable isn't working in my Rmarkdown beamer_presentation slides.

You may need to set the beamer "fragile" option, like this:

Numbers in my cells look weird!

You can change numeric formatting using number_format(). Base R options like scipen usually have no effect.

I ran caption(ht) <- "Something" and got an error message:

You may have loaded another package with a caption method, e.g. "xtable". Try loading huxtable after xtable.

How can I get line breaks in my cells?

Just insert a line break "\n" in the cell contents. Then make sure that width() is set and wrap() is TRUE (it is by default).

How can I change the font size, font etc. of captions?

There are no direct commands for this. You have to use raw HTML/TeX/other commands within the caption itself. For example to have a bold caption in HTML, you might do something like:

How do I refer to tables in bookdown?

As of version 4.3.0, this is handled automatically for you. Just set the label using label(), then in markdown text do e.g.:

I called library(huxtable) and now my data.table objects are getting printed!

Set options(huxtable.knit_print_df = FALSE).

How can I set a property on an arbitrary group of cells?

If you can't use the mapping-functions interface, and you want to set a property for multiple cells that aren't all in the same rows and/or columns, you could use a little-known fact about R subsetting. If you subset ht[x] where x is two-column numeric matrix, then each row of x indexes a single (row, column) cell. So, for example, here's how to set the background color of cells (2,1), (1, 3) and (4, 2) of a huxtable:

Another useful trick sets properties on the diagonal, using diag():

I have another problem.

If you have a bug - i.e. there is something wrong with the software - or a feature request, please report it to https://github.com/hughjonesd/huxtable/issues. Otherwise, ask a question on StackOverflow or https://community.rstudio.com. That way, other people will benefit from the answers you get.

Can I email you directly?

I'd rather you asked on a public website. If you then email me a link, I may be able to help.