knit_print.huxtable {huxtable}	R Documentation

Print a huxtable within knitr

Description

Print a huxtable within knitr

Usage

knit_print.huxtable(x, options, ...)

Arguments

x
A huxtable.

options
Not used.

...
Not used.

Details

knitr calls knitr::knit_print() on objects when they are printed in a knitr (or RMarkdown) document. The method for huxtable objects guesses the appropriate output format (including Typst documents when using the typst package) and prints itself out appropriately. You can override the output format by setting options("huxtable.knitr_output_format").

See Also

huxtable-options

Other knit_print: knit_print.data.frame()

[Package huxtable version 5.8.0 Index]