latex_float {huxtable}	R Documentation

Set the position of the table float in LaTeX

Description

Possible values include:

"h": here

"h!" definitely here

"t" top of page

"ht" here or at top of page

"b" bottom of page

"p" page of floats

Usage

latex_float(ht)

latex_float(ht) <- value

set_latex_float(ht, value)

Arguments

ht
A huxtable.

value
A string. Set to NA to reset to the default, which is "ht".

Details

See LaTeX documentation for more details.

Value

property() returns the property value(s). set_property() and map_property() return the modified huxtable.

Examples

set_latex_float(jams, "b")

[Package huxtable version 5.8.0 Index]