markdown {huxtable}	R Documentation

Interpret cell content as markdown

Description

Cells where the markdown property is TRUE will be interpreted as markdown.

Usage

markdown(ht)

markdown(ht) <- value

set_markdown(ht, row, col, value = TRUE)

map_markdown(ht, row, col, fn)

Arguments

ht
A huxtable.

value
A logical vector or matrix. Set to NA to reset to the default, which is FALSE.

row
A row specifier. See rowspecs for details.

col
An optional column specifier.

fn
A mapping function. See mapping-functions for details.

Details

Markdown is currently implemented for HTML, Word, Powerpoint, RTF, LaTeX and on-screen display. Word requires the ftExtra package.

Most formats use commonmark, with the "strikethrough" extension enabled.

The following features are intended to work:

bold and italic text

strikethrough (write ~~text~~ to strike through text).

hyperlinks

There are some quirks:

Paragraph-level properties (e.g. lists) won't work in Word.

Strikethrough will probably not work in Word.

To make lists work in LaTeX, set width() and ensure wrap() is TRUE.

Inline images in RTF work using the INCLUDEPICTURE field type.

Only local images (not urls) work in typst.

If you try to use markdown tables within a table cell, then seek psychiatric help.

Note

Markdown content in cells is completely separate from printing the whole table as markdown using print_md(). When you set markdown to TRUE, huxtable itself interprets the cell contents as markdown, and spits out HTML, TeX or whatever.

See Also

set_markdown_contents(), a shortcut function.

Examples

jams[3, 2] <- "~2.10~ **Sale!** 1.50"
set_markdown(jams, 3, 2)

[Package huxtable version 5.8.0 Index]