na_string {huxtable}	R Documentation

Change how NA values are printed

Description

NA values in the huxtable are printed as the value of na_string.

Usage

na_string(ht)

na_string(ht) <- value

set_na_string(ht, row, col, value)

map_na_string(ht, row, col, fn)

Arguments

ht
A huxtable.

value
A character vector or matrix. Set to NA to reset to the default, which is "".

row
A row specifier. See rowspecs for details.

col
An optional column specifier.

fn
A mapping function. See mapping-functions for details.

See Also

Other formatting functions: background_color(), bold(), font(), font_size(), number_format(), text_color()

Examples

jams[3, 2] <- NA
jams
set_na_string(jams, "---")

[Package huxtable version 5.8.0 Index]