position {huxtable}	R Documentation

Set the table's position with respect to surrounding content

Description

Table position may be "left", "right" or "center". If you want text to wrap around the table, use "wrapleft" or "wrapright".

Usage

position(ht)

position(ht) <- value

set_position(ht, value)

Arguments

ht
A huxtable.

value
String. "left", "center", "right", "wrapleft" or "wrapright". Set to NA to reset to the default, which is "center".

Details

"wrapleft" and "wrapright" position the table to the left or right, and allow text to wrap around the table.

Value

property() returns the property value(s). set_property() and map_property() return the modified huxtable.

Examples

set_position(jams, "left")
set_position(jams, "right")
set_position(jams, "center")

[Package huxtable version 5.8.0 Index]