print.huxtable {huxtable}	R Documentation

Format and print huxtables using a default method

Description

By default huxtables are printed using print_screen(). In certain cases, for example in Sweave documents, it may be useful to change this. You can do so by setting options("huxtable.print").

Usage

## S3 method for class 'huxtable'
print(x, ...)

## S3 method for class 'huxtable'
format(x, ..., output = c("latex", "html", "md", "screen", "rtf", "typst"))

Arguments

x
A huxtable.

...
Options passed to other methods.

output
Output format. One of "html", "latex", "md", "screen", "rtf" or "typst".

Value

print prints the huxtable and returns NULL invisibly.

format returns a string representation from to_latex(), to_html() etc.

See Also

To change how huxtables are printed within knitr, see options("huxtable.knitr_output_format") in huxtable-options

Examples

## Not run: 
# to print LaTeX output:
options(huxtable.print = print_latex)
# to print Typst output:
options(huxtable.print = print_typst)

## End(Not run)

format(jams, output = "screen")
format(jams, output = "md")
format(jams, output = "typst")

[Package huxtable version 5.8.0 Index]