restack-across-down {huxtable}	R Documentation

Restack huxtables across/down the page

Description

restack_across() splits a huxtable horizontally, then joins the parts up side by side.

restack_down() splits a huxtable vertically, then joins the parts up top to bottom.

Usage

restack_across(
  ht,
  rows,
  headers = TRUE,
  on_remainder = c("warn", "stop", "fill")
)

restack_down(
  ht,
  cols,
  headers = TRUE,
  on_remainder = c("warn", "stop", "fill")
)

Arguments

ht
A huxtable

rows, cols
How many rows/columns the new result should have.

headers
Logical. Take account of header rows/columns?

on_remainder
String. "warn", "stop" or "fill". See below.

Details

If headers is TRUE, header rows/columns will be repeated across/down the restacked huxtable as necessary.

on_remainder determines what happens if the huxtable could not be evenly divided for restacking:

"stop": stop with an error.

"fill": fill the remainder with empty cells.

"warn" (the default): issue a warning, then fill the remainder with empty cells.

Value

A new huxtable.

See Also

split-across-down

Examples

ht <- as_hux(matrix(LETTERS[1:4], 2, 2))
ht <- set_all_borders(ht)
ht

restack_down(ht, 1)
restack_across(ht, 1)

# headers:
restack_across(jams, 2)
restack_across(jams, 2,
  headers = FALSE
)

# on_remainder:
restack_across(jams, 3,
  on_remainder = "fill"
)

[Package huxtable version 5.8.0 Index]