row_height {huxtable}	R Documentation

Set the height of table rows

Description

Numeric heights are scaled to 1 and treated as proportions of the table height in HTML, or of the text height (\textheight) in LaTeX. Character row heights must be valid CSS or LaTeX dimensions.

Usage

row_height(ht)

row_height(ht) <- value

set_row_height(ht, row, value)

Arguments

ht
A huxtable.

value
Numeric or character vector. Set to NA to reset to the default, which is NA.

row
A row specifier. See rowspecs for details.

Value

row_height() returns the row_height property. set_row_height() returns the modified huxtable.

See Also

col_width(), height(), width()

Other table measurements: col_width(), height(), width()

Examples

row_height(jams) <- c(.4, .2, .2, .2)
row_height(jams)

[Package huxtable version 5.8.0 Index]