set_default_properties {huxtable}	R Documentation

Default huxtable properties

Description

Defaults are used for new huxtables, and also when a property is set to NA.

Usage

set_default_properties(...)

get_default_properties(names = NULL)

Arguments

...
Properties specified by name, or a single named list.

names
Vector of property names. If NULL, all properties are returned.

Details

Note that autoformat = TRUE in huxtable() overrides some defaults.

To set default border styles, use the pseudo-properties border/border_style/border_color. You cannot set defaults separately for different sides.

Value

For set_default_properties, a list of the previous property values, invisibly.

For get_default_properties, a list of the current defaults.

See Also

Options for autoformat in huxtable-options.

Examples

old <- set_default_properties(
  text_color = "red",
  border     = 0.4
)
hux(a = 1:2, b = 1:2)
set_default_properties(old)
get_default_properties("bold")

[Package huxtable version 5.8.0 Index]