set_markdown_contents {huxtable}	R Documentation

Set cell contents, interpreting them as markdown

Description

This convenience function calls set_contents() and set_markdown().

Usage

set_markdown_contents(ht, row, col, value)

Arguments

ht
A huxtable.

row
A row specifier. See rowspecs for details.

col
An optional column specifier.

value
Cell contents, as a markdown string.

Value

The modified huxtable.

Note

Markdown content in cells is completely separate from printing the whole table as markdown using print_md(). When you set markdown to TRUE, huxtable itself interprets the cell contents as markdown, and spits out HTML, TeX or whatever.

See Also

markdown().

Examples

set_markdown_contents(
  jams, 1, 1,
  "**Type** of jam"
)

[Package huxtable version 5.8.0 Index]