span-overwrites-shadows {huxtable}	R Documentation

Section about colspan/rowspan

Description

Section about colspan/rowspan

Cell content

In merged cell ranges, only the top left cell's content is displayed. In addition, when you merge cells (either by setting colspan() or rowspan(), or using merge_cells() and friends) the content of the top left cell is copied to other cells. This prevents unexpected changes to content if you reorder or subset rows and columns.

[Package huxtable version 5.8.0 Index]