table_environment {huxtable}	R Documentation

Set the "table" environment in LaTeX

Description

By default this is "table".

Usage

table_environment(ht)

table_environment(ht) <- value

set_table_environment(ht, value)

Arguments

ht
A huxtable.

value
A string. Set to NA to reset to the default, which is "table".

Details

No features are guaranteed to work if you set this to a non-default value. Use at your own risk! In particular, you may need to set latex_float() to a non-default value.

If position() is set to "wrapleft" or "wrapright", this value is overridden.

Value

property() returns the property value(s). set_property() and map_property() return the modified huxtable.

Examples

set_table_environment(jams, "table*")

[Package huxtable version 5.8.0 Index]