Set the table's tabular environment in LaTeX

By default this is either "tabular" or "tabularx".

tabular_environment(ht)
tabular_environment(ht) <- value
set_tabular_environment(ht, value)

Arguments

ht
A huxtable.

value
A string. Set to NA to reset to the default, which is "NA".

Value

tabular_environment() returns the tabular_environment property. set_tabular_environment() returns the modified huxtable.

Details

No features are guaranteed to work if you set this to a non-default value. Use at your own risk!

Examples

tabular_environment(jams) <-  "longtable"
tabular_environment(jams)

#> [1] "longtable"