text_color {huxtable}	R Documentation

Set the color of text in cells

Description

Colors can be in any format understood by R:

A color name like "darkred"

A HTML string like "#FF0000"

The result of a function like rgb(1, 0, 0) or grey(0.5)

Usage

text_color(ht)

text_color(ht) <- value

set_text_color(ht, row, col, value)

map_text_color(ht, row, col, fn)

Arguments

ht
A huxtable.

value
A character vector or matrix. Set to NA to reset to the default, which is NA_character_.

row
A row specifier. See rowspecs for details.

col
An optional column specifier.

fn
A mapping function. See mapping-functions for details.

See Also

Other formatting functions: background_color(), bold(), font(), font_size(), na_string(), number_format()

Examples

text_color(jams) <- "blue"
text_color(jams)

set_text_color(jams, "red")
set_text_color(jams, 2:3, 1, "red")
map_text_color(jams, by_rows("red", "blue"))

[Package huxtable version 5.8.0 Index]