print_latex {huxtable}	R Documentation

Create LaTeX representing a huxtable

Description

Create LaTeX representing a huxtable

Usage

print_latex(ht, ...)

to_latex(ht, tabular_only = FALSE, ...)

Arguments

ht
A huxtable.

...
Arguments passed to methods.

tabular_only
Return only the LaTeX tabular, not the surrounding float.

Details

If we appear to be in a rmarkdown document with the Pandoc markdown +raw_attribute extension available, to_latex will return LaTeX surrounded by a "raw attribute code block" (see https://pandoc.org/MANUAL.html#extension-raw_attribute). This helps protect against pandoc accidentally escaping the TeX code.

Value

to_latex returns a string. print_latex prints the string and returns NULL.

See Also

Other printing functions: print_html(), print_md(), print_rtf(), print_screen(), print_typst()

Examples

ht <- huxtable(
  a = 1:3,
  b = letters[1:3]
)
print_latex(ht)

[Package huxtable version 5.8.0 Index]