print_rtf {huxtable}	R Documentation

Create RTF representing a huxtable

Description

These functions print or return an RTF character string.

Usage

print_rtf(ht, fc_tables = rtf_fc_tables(ht), ...)

to_rtf(ht, fc_tables = rtf_fc_tables(ht), ...)

Arguments

ht
A huxtable.

fc_tables
See rtf_fc_tables().

...
Arguments passed to methods.

Details

RTF files use a single per-document table for colors, and one for fonts. If you are printing multiple huxtables in a document, you need to make sure that the font and color table is set up correctly and that the RTF tables refer back to them. See rtf_fc_tables().

Prepare all the huxtables;

Call rtf_fc_tables(), passing in all the huxtables;

Print the rtfFCTables object in the RTF document header;

Pass in the rtfFCTables object to each call to print_rtf.

Value

to_rtf returns a string representing an RTF table. The fc_tables attribute of the returned string will contain the fc_tables object that was passed in (or autocreated). print_rtf prints the string and returns NULL.

Limitations

rmarkdown"s rtf_document can"t yet print out customized color tables, so custom fonts and colors won"t work in this context.

col_width() and width() can only be numeric or "pt".

wrap() has no effect: cell contents always wrap.

rotation() can only be 90 or 270, i.e. text going up or down.

See Also

Other printing functions: print_html(), print_latex(), print_md(), print_screen(), print_typst()

Examples

print_rtf(jams)

[Package huxtable version 5.8.0 Index]