Print a huxtable on screen

Print a huxtable on screen

print_screen(ht, ...)

to_screen(ht, ...)

# S3 method for huxtable
to_screen(ht,
  min_width = ceiling(getOption("width")/6),
  max_width = getOption("width", Inf), compact = TRUE,
  colnames = TRUE, color = getOption("huxtable.color_screen", default =
  TRUE), ...)

Arguments

ht
A huxtable.

...
Passed on to to_screen.

min_width
Minimum width in on-screen characters of the result.

max_width
Maximum width in on-screen characters of the result. Overrides min_width.

compact
Logical. To save space, don't print lines for empty horizontal borders.

colnames
Logical. Whether or not to print colum names.

color
Logical. Whether to print the huxtable in color (requires the crayon package).

Value

to_screen returns a string. print_screen prints the string and returns NULL.

Details

colspan, rowspan, align and caption properties are shown. So are borders (but not border styles). If the crayon package is installed, output will be colorized (and contents bolded or italicized) by default; this will work in recent daily builds of RStudio as of October 2017.

See also

Examples

bottom_border(jams)[1, 1:2] <- 1
bold(jams)[1, 1:2] <- TRUE
jams <- map_text_color(jams,
      by_regex("berry" = "red"))

print_screen(jams)

#>   Type           Price  
#> ────────────────────────
#>   Strawberry      1.90  
#>   Raspberry       2.10  
#>   Plum            1.80  
#> 
#> Column names: Type, Price