valign {huxtable}	R Documentation

Set the vertical alignment of cell content

Description

Allowed values are "top", "middle", "bottom" or NA.

Usage

valign(ht)

valign(ht) <- value

set_valign(ht, row, col, value)

map_valign(ht, row, col, fn)

Arguments

ht
A huxtable.

value
A character vector or matrix. Set to NA to reset to the default, which is "top".

row
A row specifier. See rowspecs for details.

col
An optional column specifier.

fn
A mapping function. See mapping-functions for details.

Details

Vertical alignment may not work for short text in LaTeX. Defining row heights with row_height() may help.

Examples

valign(jams) <- "top"
valign(jams)

jams2 <- set_valign(jams, "bottom")
valign(jams2)

jams3 <- set_valign(jams, 2:3, 1, "bottom")
valign(jams3)

jams4 <- map_valign(jams, by_rows(
  "bottom",
  "top"
))
valign(jams4)

[Package huxtable version 5.8.0 Index]