width {huxtable}	R Documentation

Set the table width

Description

width() sets the width of the entire table, while col_width() sets the width of individual columns. A numeric width is treated as a proportion of f the surrounding block width (HTML) or text width (LaTeX). A character width must be a valid CSS or LaTeX dimension.

Usage

width(ht)

width(ht) <- value

set_width(ht, value)

Arguments

ht
A huxtable.

value
A number or string. Set to NA to reset to the default, which is NA_real_.

Value

property() returns the property value(s). set_property() and map_property() return the modified huxtable.

See Also

Other table measurements: col_width(), height(), row_height()

Examples

set_width(jams, 0.8)

[Package huxtable version 5.8.0 Index]