Convert a huxtable for Word/Powerpoint

Huxtables can be converted to flextable::flextable() objects, for use in Word and Powerpoint documents.

as_FlexTable(x, ...)

as_flextable(x, ...)

# S3 method for huxtable
as_flextable(x, colnames_to_header = FALSE, ...)

Arguments

x
A huxtable.

...
Not used.

colnames_to_header
Use huxtable column names as the header. If FALSE, the flextable will contain only a body and no header.

Value

an object of class flextable.

Details

Note: you can't use flextable Word output within rmarkdown. Instead you have to write the Word file yourself. See officer::read_docx().

as_FlexTable is deprecated and calls as_flextable with a warning.

Properties are supported, with the following exceptions:

Rotation of 0, 90 or 270 is supported.

Non-numeric column widths and row heights are not supported.

Table height, wrap, captions and table position are not supported.

Border style "double" is not supported and becomes "solid".

Challenge

Try to say as_flextable.huxtable ten times without pausing.

Examples

ht <- hux(a = 1:3, b = 1:3)
ft <- as_flextable(ht)
# NOT RUN {
  my_doc <- officer::read_docx()
  my_doc <- flextable::body_add_flextable(
        my_doc, ft)
  print(my_doc, target =
        "path/to/my_doc.docx")
# }