Convert objects to huxtables

as_huxtable or as_hux converts an object to a huxtable. Conversion methods exist for data frames, tables, ftables, matrices and (most) vectors. is_hux[table] tests if an object is a huxtable.

as_huxtable(x, ...)

as_hux(x, ...)

# S3 method for default
as_huxtable(x,
  add_colnames = getOption("huxtable.add_colnames", FALSE),
  add_rownames = FALSE, autoformat = getOption("huxtable.autoformat",
  TRUE), ...)

is_huxtable(x)

is_hux(x)

Arguments

x
Object to convert.

...
Arguments passed on to huxtable().

add_colnames
If TRUE, add a first row of column names to the huxtable.

add_rownames
If TRUE or a character string, add a first column of row names to the huxtable. The string gives the name for the new column (or "rownames" for TRUE).

autoformat
If TRUE, automatically format columns by type. See below.

Value

An object of class "huxtable".

Examples

dfr <- data.frame(
        a = 1:5,
        b = letters[1:5],
        stringsAsFactors = FALSE
      )
as_huxtable(dfr)

#>         1   a        
#>         2   b        
#>         3   c        
#>         4   d        
#>         5   e        
#> 
#> Column names: a, b

mx <- matrix(letters[1:12], 4, 3)
as_huxtable(mx)

#>   a       e       i      
#>   b       f       j      
#>   c       g       k      
#>   d       h       l      
#> 
#> Column names: V1, V2, V3

library(stats)
tbl <- table(
        Wool    = warpbreaks$wool,
        Tension = warpbreaks$tension
      )
as_huxtable(tbl) # adds row and column names by default

#>             L      M   H     
#>      A      9      9   9     
#>      B      9      9   9     
#> 
#> Column names: rownames, L, M, H

# adding rownames:
as_hux(mtcars[1:3,], add_colnames = TRUE,
      add_rownames = "Car")

#>        Car    mpg   cyl   disp    hp   drat     wt  
#>      Mazda   21       6    160   110   3.9    2.62  
#>        RX4                                          
#>      Mazda   21       6    160   110   3.9    2.88  
#>    RX4 Wag                                          
#>     Datsun   22.8     4    108    93   3.85   2.32  
#>        710                                          
#> 
#> Column names: Car, mpg, cyl, disp, hp, drat, wt, qsec, vs, am, gear, carb
#> 
#> 7/12 columns shown.