Cell text style

Functions to get or set the cell text style property of huxtable table cells.

bold(ht)
bold(ht) <- value
set_bold(ht, row, col, value, byrow = FALSE)

italic(ht)
italic(ht) <- value
set_italic(ht, row, col, value, byrow = FALSE)

Arguments

ht
A huxtable.

value
A logical vector or matrix Set to NA to reset to the default.

row
A row specifier. See rowspecs for details.

col
An optional column specifier.

byrow
If TRUE, fill in values by row rather than by column.

Value

For bold, the bold attribute. For set_bold, the ht object.

See also

Examples

ht <- huxtable(a = 1:3, b = 1:3)
bold(ht) <-  TRUE
bold(ht)

#>      a    b
#> 1 TRUE TRUE
#> 2 TRUE TRUE
#> 3 TRUE TRUE

print_screen(ht)

#>         1         1  
#>         2         2  
#>         3         3  
#> 
#> Column names: a, b

ht <- huxtable(a = 1:3, b = 3:1)
set_bold(ht, TRUE)

#>         1         3  
#>         2         2  
#>         3         1  
#> 
#> Column names: a, b

set_bold(ht, 1:2, 1, TRUE)

#>         1         3  
#>         2         2  
#>         3         1  
#> 
#> Column names: a, b

set_bold(ht, 1:2, 1:2, c(TRUE, FALSE), byrow = TRUE)

#>         1         3  
#>         2         2  
#>         3         1  
#> 
#> Column names: a, b

set_bold(ht, where(ht == 1), TRUE)

#>         1         3  
#>         2         2  
#>         3         1  
#> 
#> Column names: a, b