Map cell contents to properties using case_when

This function uses dplyr::case_when() to set cell properties.

by_cases(..., ignore_na = TRUE)

Arguments

...

A list of two-sided formulas interpreted by case_when.

ignore_na

If TRUE, NA values in the result will be left unchanged from their previous values. Otherwise, NA normally resets to the default.

Value

A function for use in map_*** functions.

Details

Within the formulas, the variable . will refer to the content of ht[rows, cols], after conversion to a vector.

case_when returns NA when no formula LHS is matched. To avoid this, set a default in the last formula: TRUE ~ default.

case_when can't deal with brdr() objects, so you cannot use these in by_cases().

See also

Examples

if (!requireNamespace("dplyr")) {
  stop("Please install the 'dplyr' package to run this example")
}

ht <- hux(runif(5), letters[1:5])

map_background_color(ht, by_cases(
  . == "a" ~ "red",
  . %in% letters ~ "green",
  . < 0.5 ~ "pink"
))
#>                                  0.29    a        
#>                                  0.678   b        
#>                                  0.735   c        
#>                                  0.196   d        
#>                                  0.981   e