Insert a row or column

These convenience functions wrap cbind or rbind for huxtables to insert a single row.

insert_column(ht, ..., after = 0, copy_cell_props = TRUE)

insert_row(ht, ..., after = 0, copy_cell_props = TRUE)

Arguments

ht
A huxtable.

...
Cell contents.

after
Insert the row/column after this position. 0 (the default) inserts as the first row/column.

copy_cell_props
Copy cell properties from the previous row or column (if after > 0). See cbind.huxtable().

Value

The modified huxtable

Details

In insert_column only, you can use a column name for after.

See also

Examples

ht <- hux(a = 1:5, b = 1:5, c = 1:5)
insert_row(ht, 2.5, 2.5, 2.5, after = 2)

#>       1       1       1  
#>       2       2       2  
#>       2       2       2  
#>       3       3       3  
#>       4       4       4  
#>       5       5       5  
#> 
#> Column names: a, b, c

insert_column(ht, 5:1)

#>      5      1      1      1  
#>      4      2      2      2  
#>      3      3      3      3  
#>      2      4      4      4  
#>      1      5      5      5  
#> 
#> Column names: V1, a, b, c

insert_column(ht, 5:1, after = 3)

#>      1      1      1      5  
#>      2      2      2      4  
#>      3      3      3      3  
#>      4      4      4      2  
#>      5      5      5      1  
#> 
#> Column names: a, b, c, V1

insert_column(ht, 5:1, after = "b")

#>      1      1      5      1  
#>      2      2      4      2  
#>      3      3      3      3  
#>      4      4      2      4  
#>      5      5      1      5  
#> 
#> Column names: a, b, V1, c