label {huxtable}	R Documentation

Set a table label for external referencing

Description

The label is used as the table's label in LaTeX, and as the "id" property of the table element in HTML.

Usage

label(ht)

label(ht) <- value

set_label(ht, value)

Arguments

ht
A huxtable.

value
A string. Set to NA to reset to the default, which is NA_character_.

Details

LaTeX table labels typically start with "tab:".

Within knitr, huxtable labels will default to the same as the knitr chunk label. To turn off this behaviour, set options(huxtable.autolabel = FALSE).

If you use bookdown, and set a label on your table, the table caption() will automatically be prefixed with (#label). You can then refer to the table using @ref(label). label needs to start with "tab:"; if it doesn't, the "tab:" prefix will be added automatically. To turn off this behaviour, set options(huxtable.bookdown = FALSE).

Value

property() returns the property value(s). set_property() and map_property() return the modified huxtable.

See Also

huxtable-options

Examples

set_label(jams, "tab:mytable")

[Package huxtable version 5.8.0 Index]