Change how NA values are printed

NA values in the huxtable are printed as the value of na_string.

Usage

na_string(ht)

na_string(ht) <- value

set_na_string(ht, row, col, value)

map_na_string(ht, row, col, fn)

Arguments

ht

A huxtable.

value

A character vector or matrix. Set to NA to reset to the default, which is "".

row

A row specifier. See rowspecs for details.

col

An optional column specifier.

fn

A mapping function. See mapping-functions for details.

See also

Examples

jams[3, 2] <- NA
jams
#>                               Type           Price  
#>                               Strawberry      1.90  
#>                               Raspberry             
#>                               Plum            1.80  
#> 
#> Column names: Type, Price
set_na_string(jams, "---")
#>                               Type           Price  
#>                               Strawberry      1.90  
#>                               Raspberry     ---     
#>                               Plum            1.80  
#> 
#> Column names: Type, Price