Override a model's tidy output

Use tidy_override to provide your own p values, confidence intervals etc. for a model.

tidy_override(x, ..., glance = list(), extend = FALSE)

# S3 method for tidy_override
tidy(x, ...)

# S3 method for tidy_override
glance(x, ...)

# S3 method for tidy_override
nobs(object, ...)

Arguments

x
A model with methods defined for generics::tidy() and/or generics::glance().

...
In tidy_override, columns of statistics to replace tidy output. In tidy and glance methods, arguments passed on to the underlying model.

glance
A list of summary statistics for glance.

extend
Logical: allow adding new statistics?

object
A tidy_override object.

Value

An object of class "tidy_override". When tidy and glance are called on this, it will return results from the underlying model, replacing some columns with your own data.

Examples

if (! requireNamespace("broom")) {
  stop("Please install 'broom' to run this example.")
}

lm1 <- lm(mpg ~ cyl, mtcars)
fixed_lm1 <- tidy_override(lm1,
      p.value = c(.04, .12),
      glance = list(r.squared = 0.99))

broom::tidy(fixed_lm1)

#> # A tibble: 2 x 5
#>   term        estimate std.error statistic p.value
#>   <chr>          <dbl>     <dbl>     <dbl>   <dbl>
#> 1 (Intercept)    37.9      2.07      18.3     0.04
#> 2 cyl            -2.88     0.322     -8.92    0.12

cbind(huxreg(fixed_lm1), huxreg(lm1))

#> ─────────────────────────────────────────────────────────────────────────────
#>                            (1)                                   (1)         
#>                    ────────────────────                  ────────────────────
#>   (Intercept)                37.885 *   (Intercept)              37.885 ***  
#>                              (2.074)                             (2.074)     
#>   cyl                        -2.876     cyl                      -2.876 ***  
#>                              (0.322)                             (0.322)     
#>                    ────────────────────                  ────────────────────
#>   N                          32         N                        32          
#>   R2                          0.990     R2                        0.726      
#>   logLik                    -81.653     logLik                  -81.653      
#>   AIC                       169.306     AIC                     169.306      
#> ─────────────────────────────────────────────────────────────────────────────
#>   *** p < 0.001; ** p < 0.01; * p       *** p < 0.001; ** p < 0.01; * p      
#>   < 0.05.                               < 0.05.                              
#> 
#> Column names: names, model1, names, model1